fix nvm install and update condition for style os

+++ .setup/options/packages/npm.sh
@@ -1,18 +1,37 @@
 #!/bin/bash
 
-COMMAND_NAME="n"
-if ! command -v $COMMAND_NAME &>/dev/null; then
+echo "=========================== nvm ==========================="
+COMMAND_NAME="nvm"
+if [ -s "$HOME/.nvm/nvm.sh" ]; then
     echo "$COMMAND_NAME could not be found. Setting up $COMMAND_NAME."
-    sudo npm cache clean -f
-    sudo npm install -g $COMMAND_NAME stable
+    curl https://raw.githubusercontent.com/nvm-sh/nvm/master/install.sh | bash
+
+    source "$HOME/.profile"
+
+    nvm install node
+    nvm use node
+    nvm alias default node
 else
     echo "$COMMAND_NAME install ok installed"
 fi
+echo ""
 
-COMMAND_NAME="npm"
-if ! command -v $COMMAND_NAME &>/dev/null; then
-    echo "$COMMAND_NAME could not be found. Setting up $COMMAND_NAME."
-    sudo npm install -g $COMMAND_NAME
-else
-    echo "$COMMAND_NAME install ok installed"
+# if nvm is not installed
+if [ ! -s "$HOME/.nvm/nvm.sh" ]; then
+    COMMAND_NAME="n"
+    if ! command -v $COMMAND_NAME &>/dev/null; then
+        echo "$COMMAND_NAME could not be found. Setting up $COMMAND_NAME."
+        sudo npm cache clean -f
+        sudo npm install -g $COMMAND_NAME stable
+    else
+        echo "$COMMAND_NAME install ok installed"
+    fi
+
+    COMMAND_NAME="npm"
+    if ! command -v $COMMAND_NAME &>/dev/null; then
+        echo "$COMMAND_NAME could not be found. Setting up $COMMAND_NAME."
+        sudo npm install -g $COMMAND_NAME
+    else
+        echo "$COMMAND_NAME install ok installed"
+    fi
 fi
